requirements.txt
fastapi[standard]
PyJWT
pymongo
python-dotenv
aiohttp